requirements.txt
pandas>=2.0
numpy>=1.24
pyarrow>=12.0
pyyaml>=6.0
requests>=2.31
tqdm>=4.65
scikit-learn>=1.3
scipy>=1.11
matplotlib>=3.7
seaborn>=0.12

# SFT / LoRA
torch>=2.1
transformers>=4.40
peft>=0.10
trl>=0.8
unsloth>=2024.3
datasets>=2.18
bitsandbytes>=0.43
accelerate>=0.28

# Evaluation
rouge-score>=0.1.2
sentence-transformers>=2.6
